+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-20T13:56:36.490Z",
+    "2026-07-20T15:39:47.959Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {