+++ packages/orq-rc/RELEASES.md
@@ -10959,3 +10959,13 @@ Based on:
 - [typescript v4.12.0-rc.42] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.42] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.42 - packages/orq-rc
+
+## 2026-07-20 15:43:06
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.2 (2.918.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.43] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.43] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.43 - packages/orq-rc